Update 029_make_journald_persistent.md
man journald.conf 
"Storage= Controls where to store journal data. ... If "persistent", data will be stored preferably on disk, i.e. below the /var/log/journal hierarchy (which is created if needed), with a fallback to /run/log/journal (which is created if needed), during early boot and if the disk is not writable."

So just editing the file /etc/systemd/journald.conf and changing "#Storage=auto" to "Storage=persistent" works without having to memorize which directory has to be created, which is a common error. The directory is created by the service automatically once "persistent" is selected.

I also added ".service" to the systemctl command because it is more accurate this way.