Update 008_create_users_with_specified_uid.md
Switched the existing working command for a command that can be copied from the man pages and that allows users to just paste the amount of days they wish to have added until account expiration. Copy-paste means less room for error and just adding the amount of days takes away the risk of typing the date out wrong or making a mistake when adding a month to today's date.